+++ firmware/src/modules/screen.rs
@@ -10,11 +10,13 @@ use embedded_dma::Word;
 use embedded_graphics::{
     pixelcolor::{raw::RawU16, Bgr565, Gray4},
     prelude::*,
+    primitives::{PrimitiveStyle, Rectangle},
     text::{Alignment, Baseline, Text, TextStyle, TextStyleBuilder},
 };
 use embedded_graphics_framebuf::{backends::FrameBufferBackend, FrameBuf};
 use embedded_hal::timer::CountDown as _;
 use jukebox_util::{
+    peripheral::Connection,
     screen::{ProfileName, ScreenProfile},
     stats::SystemStats,
 };
@@ -28,12 +30,13 @@ use rp2040_hal::{
     timer::CountDown,
     Timer,
 };
+use usb_device::device::UsbDeviceState;
 
 use crate::{
     st7789::St7789,
     util::{
-        DEFAULT_PROFILE_NAME, DEFAULT_SCREEN_PROFILE, DEFAULT_SYSTEM_STATS, ICONS, PROFILE_NAME,
-        SCREEN_CONTROLS, SCREEN_SYSTEM_STATS,
+        CONNECTION_STATUS, DEFAULT_PROFILE_NAME, DEFAULT_SCREEN_PROFILE, DEFAULT_SYSTEM_STATS,
+        ICONS, PROFILE_NAME, SCREEN_CONTROLS, SCREEN_SYSTEM_STATS, USB_STATUS,
     },
 };
 
@@ -99,8 +102,16 @@ static FONT1: BitmapFont<'static, Gray4, 1> = mplus!(
     'a'..'z',
     [" ", "-", ".", "%", "°", "/", ":"]
 );
-static FONT2: BitmapFont<'static, Gray4, 1> =
-    mplus!(code(100), 500, 32, false, 1, 4, '0'..='9', [".", " "]);
+static FONT2: BitmapFont<'static, Gray4, 1> = mplus!(
+    code(100),
+    500,
+    32,
+    false,
+    1,
+    4,
+    '0'..='9',
+    ["A", "N", ".", " ", "/"]
+);
 
 const LEFT_TEXT_STYLE: TextStyle = TextStyleBuilder::new()
     .alignment(Alignment::Left)
@@ -123,6 +134,7 @@ pub struct ScreenMod {
     profile_name: ProfileName,
     screen_profile: ScreenProfile,
     system_stats: SystemStats,
+    usb_state: UsbDeviceState,
 
     dma_ch0: Channel<CH0>,
     timer: CountDown,
@@ -148,6 +160,7 @@ impl ScreenMod {
             profile_name: DEFAULT_PROFILE_NAME,
             screen_profile: DEFAULT_SCREEN_PROFILE,
             system_stats: DEFAULT_SYSTEM_STATS,
+    usb_state: UsbDeviceState::Default,
 
             dma_ch0,
             timer,
@@ -233,9 +246,7 @@ impl ScreenMod {
         }
     }
 
-    fn draw_pre_tick(&mut self) {
-        // TODO: brightness control? via pwm?
-
+    fn draw_pre_tick(&mut self, uid: &'static str, ver: &'static str) {
         let c: Bgr565 = {
             let c: RawU16 = self.screen_profile.text_color().into();
             c.into()
@@ -261,13 +272,48 @@ impl ScreenMod {
                     .font(&FONT1)
                     .build();
 
-                let _ = Text::with_text_style(
-                    self.profile_name.to_str(),
-                    Point::new(160 - 1, 224),
-                    font1_style.clone(),
-                    CENTER_TEXT_STYLE,
-                )
-                .draw(&mut self.fb);
+                let mut serial_status = Connection::NotConnected(false);
+                CONNECTION_STATUS.with_lock(|c| {
+                    serial_status = *c;
+                });
+
+                if serial_status != Connection::Connected {
+                    let _ = Text::with_text_style(
+                        uid,
+                        Point::new(160 - 1, 224),
+                        font1_style.clone(),
+                        CENTER_TEXT_STYLE,
+                    )
+                    .draw(&mut self.fb);
+
+                    let _ = Text::with_text_style(
+                        ver,
+                        Point::new(255, 224),
+                        font1_style.clone(),
+                        LEFT_TEXT_STYLE,
+                    )
+                    .draw(&mut self.fb);
+
+                    let usb_color = if self.usb_state != UsbDeviceState::Configured {
+                        Bgr565::new(8, 16, 31)
+                    } else {
+                        Bgr565::new(8, 63, 8)
+                    };
+
+                    let _ = Rectangle::new(Point::new(26, 224 - 10), Size::new(16, 16))
+                        .into_styled(PrimitiveStyle::with_fill(usb_color))
+                        .draw(&mut self.fb);
+
+                    self.rounded_rect(bgc, 26, 224 - 10, 16, 1);
+                } else {
+                    let _ = Text::with_text_style(
+                        self.profile_name.to_str(),
+                        Point::new(160 - 1, 224),
+                        font1_style.clone(),
+                        CENTER_TEXT_STYLE,
+                    )
+                    .draw(&mut self.fb);
+                }
             }
             ScreenProfile::DisplayStats {
                 brightness: _,
@@ -531,7 +577,13 @@ impl ScreenMod {
         }
     }
 
-    pub fn update(mut self, keys: &[bool], _t: &Timer) -> Self {
+    pub fn update(
+        mut self,
+        keys: &[bool],
+        _t: &Timer,
+        uid: &'static str,
+        ver: &'static str,
+    ) -> Self {
         for i in 0..12 {
             if keys[i] {
                 self.keys_status[i] = 4;
@@ -560,6 +612,13 @@ impl ScreenMod {
             }
             p.0 = false;
         });
+        USB_STATUS.with_mut_lock(|p| {
+            if p.0 {
+                changed = true;
+                self.usb_state = p.1;
+            }
+            p.0 = false;
+        });
 
         if changed {
             // using multiple channels did not meaningfully improve performance.
@@ -578,8 +637,7 @@ impl ScreenMod {
                 dma_ch0
             };
 
-            self.draw_pre_tick();
-
+            self.draw_pre_tick(uid, ver);
             self.keys_previous_frame = [2; 12];
             self.draw_post_tick();
         }